purgevieilles: mode description
= On affiche juste ce qui pourrait être fait, mais on ne le fait pas.

darcs-hash:8207b7a5edcd68e5f43d208f3090289ad7c37d21

+++ scripts/sauvegarde/purgevieilles.php
@@ -14,11 +14,15 @@ $periodes = array
 
 $fichiers = "/Users/gui/Downloads/Mail.*.tar.gz";
 
+$faire = true;
 array_shift($argv);
 while(($arg = array_shift($argv)) !== null)
 {
 	switch($arg)
 	{
+		case '-n':
+			$faire = false;
+			break;
 		default:
 			$fichiers = $arg;
 			break;
@@ -55,10 +59,15 @@ foreach($t as $d => $fichier)
 		{
 			$dGarde = $d;
 			$garde = $fichier;
+			if(!$faire)
+				echo '+ '.$fichier."\n";
 			continue;
 		}
 	
-	unlink($fichier);
+	if(!$faire)
+		echo '- '.$fichier."\n";
+	else
+		unlink($fichier);
 }
 
 ?>
